# specs/changes/active/20260912-iam-identity-access-management/contracts/internal-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IAM资源内部授权契约",
    "version": "2.0.0"
  },
  "paths": {
...
    "/inner/iam/field-bindings/manifest": {
      "post": {
        "operationId": "fieldBindingManifest",
        "description": "固定服务白名单及专用HMAC密钥；固定purpose与调用服务；不需要用户身份，不包含业务原值",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedResourceObjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "R信封；精确错误码区分拒绝和不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFieldBindingManifest"
                }
              }
            }
          },
          "400": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Binding": {
        "type": "object",
        "properties": {
          "actionCode": {
            "type": "string"
          },
          "use": {
            "type": "string",
            "enum": [
              "READ",
              "WRITE",
              "FILTER"
            ]
          },
          "valueType": {
            "type": "string"
          },
          "property": {
            "type": "string"
          },
          "fieldKey": {
            "type": "string"
          },
          "textual": {
            "type": "boolean"
          }
        }
      },
...
      "FieldBindingManifest": {
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceKey"
          },
          "version": {
            "type": "string"
          },
          "bindings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Binding"
            }
          }
        }
      },
...
      "IamErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "success"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "InvalidArgument",
              "IdentityInvalid",
              "ActionDenied",
              "PasswordChangeRequired",
              "DataScopeDenied",
              "DelegationExceeded",
              "RoleRevisionUnavailable",
              "ApplicationUnavailable",
              "ObjectNotFound",
              "PolicyConflict",
              "RevisionConflict",
              "ObjectInUse",
              "StepUpFailed",
              "AuthorizationUnavailable"
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "enum": [
              null
            ]
          },
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          }
        }
      },
...
      "RFieldBindingManifest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/FieldBindingManifest"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ResourceKey": {
        "required": [
          "applicationCode",
          "domain",
          "resourceCode"
        ],
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "enum": [
              "PLATFORM",
              "TENANT"
            ]
          },
          "applicationCode": {
            "type": "string"
          },
          "resourceCode": {
            "type": "string"
          }
        }
      },
...
      "SignedResourceObjectRequest": {
        "type": "object",
        "properties": {
          "payload": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "signature": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}